+++ .husky/pre-commit
@@ -3,6 +3,7 @@
 
 # validate runs test:e2e:run won't run if port 3000 is already in use
 if lsof -Pi :3000 -sTCP:LISTEN -t >/dev/null ; then
+    echo "🚨🚨🚨"
     echo "Port 3000 is already in use, please close the process and try again."
     exit 1
 else
