+++ app/components/templates/layout/index.ts
@@ -1,1 +1,2 @@
+export * from './page-footer.tsx'
 export * from './page-header.tsx'
